test: fix & separate tests
