+++ tests/integration/test_charm_base_image.py
@@ -1,0 +1,56 @@
+# Copyright 2024 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+"""Integration tests for github-runner charm containing one runner."""
+
+from github.Branch import Branch
+from github.Repository import Repository
+from github.WorkflowRun import WorkflowRun
+from juju.application import Application
+from juju.model import Model
+
+from charm_state import BASE_IMAGE_CONFIG_NAME
+from tests.integration.helpers import (
+    DISPATCH_E2E_TEST_RUN_WORKFLOW_FILENAME,
+    dispatch_workflow,
+    ensure_charm_has_runner,
+    get_runner_name,
+    run_in_lxd_instance,
+)
+
+
+async def test_runner_base_image(
+    model: Model,
+    app_no_runner: Application,
+    github_repository: Repository,
+    test_github_branch: Branch,
+) -> None:
+    """
+    arrange: A runner with noble as base image.
+    act: Dispatch a workflow.
+    assert: A runner is created with noble OS base and the workflow job is successfully run.
+    """
+    await app_no_runner.set_config(
+        {
+            BASE_IMAGE_CONFIG_NAME: "noble",
+        }
+    )
+    await ensure_charm_has_runner(app_no_runner, model)
+
+    unit = app_no_runner.units[0]
+    runner_name = await get_runner_name(unit)
+    code, stdout, stderr = await run_in_lxd_instance(unit, runner_name, "lsb_release -a")
+    assert code == 0, f"Unable to get release name, {stdout} {stderr}"
+    assert "noble" in str(stdout)
+
+    workflow = await dispatch_workflow(
+        app=app_no_runner,
+        branch=test_github_branch,
+        github_repository=github_repository,
+        conclusion="success",
+        workflow_id_or_name=DISPATCH_E2E_TEST_RUN_WORKFLOW_FILENAME,
+        dispatch_input={"runner-tag": app_no_runner.name},
+    )
+
+    workflow_run: WorkflowRun = workflow.get_runs()[0]
+    assert workflow_run.status == "success"
